Switch Rose over to unsized List
I thought that Rose needed sized List but it seems that it doesn't.
Given that this case was the primary motivation to add sizes to the
inductive types in the first place, maybe it's not even worth keeping
sized List (or Vec) around anymore. Rose still needs to be sized in
order to pass the positivity checker though, at least when presented in
this inductive fashion instead of using containers or IR codes.

@jonsterling